+++ src/helpers.py
@@ -1,32 +1,5 @@
 
 ## Functions I seem to use over and over
 
-from canvasapi import Canvas
-import getpass
-import sys
-import pandas as pd
-import re
-import ast
-from interface import shut_down
-
-
-def _create_csv(df, output_name):
-    print(df.head())
-    while True:
-        confirmation = input("Your csv will be called: {}\nDo you want to generate this csv from with the data above? (y/n): ".format(output_name))
-        
-        if confirmation.upper() == "Y":
-            df.to_csv(output_name, index=False)
-            print("\n{} created.\nBye!".format(output_name))
-            break
-        elif confirmation.upper() =="N":
-            shut_down("\nCsv not created. You can run the script again or exit for no further action.\n")
-            break
-        else:
-            print("Please enter 'Y' to accept or 'N' to exit\n")
-            continue
-
-
-
