+++ src/interface.py
@@ -1,10 +1,3 @@
-"""
-GROUP CSV: interface & utl
-authors:
-@markoprodanovic
-last edit:
-September 24, 2020
-"""
 
 import getpass
 from canvasapi import Canvas
@@ -69,3 +62,19 @@ def _prompt_for_confirmation(user_name, course_name):
 
     confirm_strict('Would you like to continue using the above information?')
         
+
+def _create_csv(df, output_name):
+    print(df.head())
+    while True:
+        confirmation = input("Your csv will be called: {}\nDo you want to generate this csv from with the data above? (y/n): ".format(output_name))
+        
+        if confirmation.upper() == "Y":
+            df.to_csv(output_name, index=False)
+            print("\n{} created.\nBye!".format(output_name))
+            break
+        elif confirmation.upper() =="N":
+            shut_down("\nCsv not created. You can run the script again or exit for no further action.\n")
+            break
+        else:
+            print("Please enter 'Y' to accept or 'N' to exit\n")
+            continue
